Extract_text_from_video_files/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
audioread==3.0.1
azure-cognitiveservices-speech==1.40.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
datasets==3.0.0
decorator==5.1.1
dill==0.3.8
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.6.1
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-speech==2.27.0
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
huggingface-hub==0.25.0
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.3
numba==0.60.0
numpy==2.0.2
openai-whisper==20231117
packaging==24.1
pandas==2.2.3
platformdirs==4.3.6
pocketsphinx==5.0.3
pooch==1.8.2
proto-plus==1.24.0
protobuf==5.28.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydub==0.25.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sounddevice==0.5.0
soundfile==0.12.1
soxr==0.5.0.post1
SpeechRecognition==3.10.4
srt==3.5.3
sympy==1.13.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.4.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
vosk==0.3.45
websockets==13.1
wget==3.2
xxhash==3.5.0
yarl==1.11.1
